// types/profile.ts
export type Profile = {
  id: string
  full_name: string | null
  first_name: string | null
  last_name: string | null
  phone: string | null
  location: string | null
  membership_plan: string | null
  plan_price: string | null
  next_billing_date: string | null
  payment_last4: string | null
  member_since_months: string | null
  stripe_customer_id: string | null
  subscription_status: string | null
  subscription_price_id: string | null
  current_period_end: string | null
}